More long types in telegram

+++ backend/src/main/java/de/njsm/versusvirus/backend/telegram/TelegramApiWrapper.java
@@ -46,7 +46,7 @@ class TelegramApiWrapper {
         executeQuery(call);
     }
 
-    public void deleteMessage(int chatId, int messageId) {
+    public void deleteMessage(long chatId, long messageId) {
         LOG.debug("Deleting message in chat {}", chatId);
         Call<TelegramResponse<Void>> call = apiClient.deleteMessage(chatId, messageId);
         executeQuery(call);
